+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Release 0.14.4 (unreleased)
 ====================================
 
 * Fix SVN commands raising ``UnicodeDecodeError`` on non-UTF-8 output (#1383)
+* Fix a gitlink with no ``.gitmodules`` entry aborting fetches outside ``src`` (#1380)
 
 Release 0.14.3 (released 2026-06-25)
 ====================================